+++ src/components/MainNav/Submenus/Community/index.tsx
@@ -7,7 +7,7 @@ import CallToAction from '../CallToAction'
 import { Block, TwoCol, Wrapper } from '../Wrapper'
 import { graphql, navigate, useStaticQuery } from 'gatsby'
 import slugify from 'slugify'
-import { Avatar, Login } from '../../../../pages/community/Dashboard'
+import { Avatar, Login } from '../../../../pages/community/dashboard'
 import { useUser } from 'hooks/useUser'
 import getAvatarURL from '../../../Squeak/util/getAvatar'
 import { AbTesting, Analytics, FeatureFlags, SessionRecording } from 'components/ProductIcons'
