+++ src/pages/questions/index.tsx
@@ -1,12 +1,9 @@
 import React, { useEffect, useState } from 'react'
 import qs from 'qs'
-import Layout from 'components/Layout'
-import { SEO } from 'components/seo'
-import community from 'sidebars/community.json'
-import PostLayout from 'components/PostLayout'
 import SidebarSearchBox from 'components/Search/SidebarSearchBox'
 import QuestionForm from 'components/Questions/QuestionForm'
 import TopicsTable from 'components/Questions/TopicsTable'
+import CommunityLayout from 'components/Community/Layout'
 
 const fetchTopicGroups = async () => {
     // FIXME: This is has to fetch _every_ (or probably at most 25) quesiton that's part of a topic even though we only need the most recent one
@@ -46,39 +43,35 @@ export default function Questions() {
     }, [])
 
     return (
-        <Layout>
-            <PostLayout sidebar={<>Sidebar!</>} title={'Questions'} menu={community} hideSurvey>
-                <SEO title={'Questions - PostHog'} />
-
-                <div className="max-w-6xl mx-auto space-y-8 pb-12">
-                    <section className="max-w-6xl mx-auto">
-                        <div className="w-full sm:flex items-center mb-8">
-                            <h1 className="text-4xl m-0">Community questions</h1>
-                            <div className="ml-auto sm:mt-0 mt-4">
-                                <QuestionForm onSubmit={() => null} />
-                            </div>
+        <CommunityLayout title="Questions">
+            <div className="max-w-6xl mx-auto space-y-8 pb-12">
+                <section className="max-w-6xl mx-auto">
+                    <div className="w-full sm:flex items-center mb-8">
+                        <h1 className="text-4xl m-0">Community questions</h1>
+                        <div className="ml-auto sm:mt-0 mt-4">
+                            <QuestionForm onSubmit={() => null} />
                         </div>
+                    </div>
 
-                        <div className="full">
-                            <SidebarSearchBox filter="question" />
+                    <div className="full">
+                        <SidebarSearchBox filter="question" />
+                    </div>
+
+                    {topicGroups?.length > 0 && (
+                        <div className="mt-8 flex flex-col space-y-8">
+                            {topicGroups
+                                .sort(
+                                    (a, b) =>
+                                        topicGroupsSorted.indexOf(a?.attributes?.label) -
+                                        topicGroupsSorted.indexOf(b?.attributes?.label)
+                                )
+                                .map(({ attributes: { label, topics } }) => {
+                                    return <TopicsTable key={label} topicGroup={label} topics={topics} />
+                                })}
                         </div>
-
-                        {topicGroups?.length > 0 && (
-                            <div className="mt-8 flex flex-col space-y-8">
-                                {topicGroups
-                                    .sort(
-                                        (a, b) =>
-                                            topicGroupsSorted.indexOf(a?.attributes?.label) -
-                                            topicGroupsSorted.indexOf(b?.attributes?.label)
-                                    )
-                                    .map(({ attributes: { label, topics } }) => {
-                                        return <TopicsTable key={label} topicGroup={label} topics={topics} />
-                                    })}
-                            </div>
-                        )}
-                    </section>
-                </div>
-            </PostLayout>
-        </Layout>
+                    )}
+                </section>
+            </div>
+        </CommunityLayout>
     )
 }
